+++ src/main/java/com/example/housestayspringboot/service/AdminService.java
@@ -4,5 +4,5 @@ import com.example.housestayspringboot.entity.Admin;
 
 public interface AdminService {
     Admin login(String username, String password);
-    Admin findById(Long adminId);
+    Admin findById(Integer adminId);
 }
